feat: improve css / changing purple / make uploads page server side / create page 404
