--- src/components/Button/Button.module.css
@@ -1,209 +1,0 @@
-.linkButton {
-    padding: 1rem;
-    box-shadow: var(--box-shadow);
-    border-radius: var(--border-radius);
-    cursor: pointer;
-    margin-bottom: 2rem;
-    text-align: center;
-    color: var(--purple);
-    font-weight: bold;
-    display: flex;
-    align-items: center;
-    font-size: 1.3rem;
-    transition: box-shadow 0.3s ease, transform 0.3s ease;
-}
-
-.backButton {
-    position: fixed;
-    top: 6rem;
-    left: 6rem;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    border: none;
-    cursor: pointer;
-    z-index: 1000;
-    font-weight: bold;
-    background-color: transparent;
-    font-size: 1.2rem;
-    padding: 1rem;
-    color: var(--purple);
-    font-family: "Afacad Flux", sans-serif;
-    transition: box-shadow 0.2s ease, transform 0.2s ease;
-    border: none;
-}
-
-.backButton:hover {
-    transform: translateY(-4px);
-}
-
-.backButton>.icon {
-    width: 1rem;
-    height: 1rem;
-}
-
-.div {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    margin-bottom: 1rem;
-}
-
-.link {
-    width: 100%;
-    display: flex;
-    align-items: center;
-    padding: 0.5rem 2rem;
-    border-radius: var(--border-radius);
-    border: none;
-    cursor: pointer;
-    background-color: transparent;
-}
-
-.link:hover {
-    color: var(--white);
-}
-
-.linkContainer {
-    background-color: var(--light-gray);
-    box-shadow: var(--box-shadow);
-    border-radius: var(--border-radius);
-    text-align: center;
-    color: var(--purple);
-    font-weight: bold;
-    width: 95%;
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    font-size: 1.3rem;
-    transition: box-shadow 0.3s ease, transform 0.3s ease;
-    padding: 0.5rem;
-}
-
-.linkContainer:hover {
-    background-color: var(--purple-hover);
-    color: var(--white);
-    transform: translateY(-4px);
-}
-
-.linkContainer:hover>.icon {
-    color: var(--white);
-}
-
-.linkContainer .pdfLabel {
-    overflow: hidden;
-    white-space: nowrap;
-    text-overflow: ellipsis;
-    max-width: 90%;
-}
-
-
-.deleteButton {
-    border: none;
-    background-color: transparent;
-    transition: box-shadow 0.3s ease, transform 0.3s ease;
-    cursor: pointer;
-}
-
-.deleteButton:hover {
-    border: none;
-    background-color: transparent;
-    transform: translateY(-4px);
-}
-
-.trashIcon {
-    width: 1.5rem;
-    height: 1.5rem;
-    color: var(--purple);
-}
-
-.trashIcon:hover {
-    color: red;
-}
-
-.menuButton {
-    margin: 1rem 0 2rem 0;
-    padding: 2rem 1.5rem;
-    background-color: var(--light-gray);
-    box-shadow: var(--box-shadow);
-    border-radius: var(--border-radius);
-    cursor: pointer;
-    color: var(--purple);
-    font-weight: bold;
-    display: flex;
-    text-align: center;
-    max-width: 50%;
-    border: none;
-    align-items: center;
-    font-size: 1.3rem;
-    transition: box-shadow 0.3s ease, transform 0.3s ease;
-}
-
-.menuButton:hover {
-    background-color: var(--purple-hover);
-    color: var(--white);
-    transform: translateY(-4px);
-}
-
-.button {
-    padding: 0.8rem;
-    background-color: var(--purple);
-    box-shadow: var(--box-shadow);
-    border-radius: var(--border-radius);
-    cursor: pointer;
-    color: var(--white);
-    cursor: pointer;
-    font-weight: bold;
-    display: flex;
-    border: none;
-    font-size: 1.2rem;
-    transition: box-shadow 0.3s ease, transform 0.1s ease;
-    font-family: "Afacad Flux", sans-serif;
-}
-
-.button:hover {
-    background-color: var(--light-gray);
-    color: var(--teal);
-    transform: translateY(-2px);
-}
-
-.icon {
-    width: 2rem;
-    height: 2rem;
-    margin-right: 1rem;
-    color: var(--teal);
-}
-
-@media (max-width: 768px) {
-    .backButton {
-        display: flex;
-    }
-
-    .menuButton {
-        margin: 1rem;
-        padding: 2rem;
-        max-width: 100%;
-    }
-
-    .trashIcon {
-        width: 1.2rem;
-        height: 1.2rem;
-    }
-
-    .link {
-        padding: 0.5rem;
-    }
-
-    .linkContainer {
-        width: 90%;
-    }
-
-    .backButton {
-        left: 1rem
-    }
-
-    .linkContainer .pdfLabel {
-        max-width: 60%;
-    }
-
-}
