+++ src/components/Container/Container.module.css
@@ -3,7 +3,7 @@
     display: flex;
     justify-content: center;
     align-items: center;
-    padding: 2rem;
+    padding: 1.5rem;
     width: 80%;
     height: 80vh;
     margin: 3rem auto;
@@ -19,6 +19,8 @@
     height: 100%;
     width: 100%;
     display: flex;
+    justify-content:center ;
+
 }
 
 @media (max-width: 768px) {
