+++ src/components/Container/Container.tsx
@@ -1,4 +1,3 @@
-'use client';
 import React, { ReactNode } from 'react';
 import styles from "./Container.module.css";
