+++ src/components/PdfList/PdfList.module.css
@@ -1,0 +1,38 @@
+
+.pdfsDiv {
+    margin: 0 auto;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    width: 80%;
+}
+
+.pdfRow {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    width: 100%;
+}
+
+.pdfButton {
+    display: flex;
+    justify-content: flex-start;
+    width: 90%;
+    overflow: hidden;
+    text-overflow: ellipsis;
+    white-space: nowrap;
+}
+
+.deleteButton {
+    width: 10%;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
+
+@media (max-width: 768px) {
+    .pdfsDiv {
+        width: 100%;
+    }
+}
